Extract Pref Dialog utilities
