+++ package.json
@@ -57,7 +57,6 @@
   },
   "dependencies": {
     "autobind-decorator": "1.3.4",
-    "lodash": "^4.17.2",
     "mobx": "2.4.1",
     "mobx-react": "3.5.3",
     "mobx-react-devtools": "^4.2.11",
