+++ src/util/common.js
@@ -22,3 +22,17 @@ export function toTruncFixed(number, place) {
 export function toStringWithCommas(number) {
   return number.toString().replace(/\B(?=(?:\d{3})+(?!\d))/g, ',');
 };
+
+/**
+ * Parse a string if it's an integer, float, or just a string
+ */
+export function parseStringIfNumber(str) {
+  if (parseInt(str, 10).toString() === str) {
+    return parseInt(str, 10);
+  }
+  if (parseFloat(str).toString() === str) {
+    return parseFloat(str);
+  }
+
+  return str;
+}
